Move some input handling out of TopLevel
